int para double

+++ Source Code Inspection/src/br/calebe/ticketmachine/core/Troco.java
@@ -18,7 +18,7 @@ class Troco {
         this.papeisMoeda = papeisMoeda;
     }
 
-    public Troco(int valor) {
+    public Troco(double valor) {
         papeisMoeda = new PapelMoeda[6];
         int count = 0;
         while (valor % 100 != 0) {
